Authentication module completed. User controls for Dictiona app for registration completed. Client and subscription managers to be continued.

+++ Master Diction/Diction Master - Library/Component.cs
@@ -11,7 +11,7 @@ namespace Diction_Master___Library
     [XmlInclude(typeof(CompositeComponent))]
     public abstract class Component
     {
-        public int ID { get; set; }
-        public int ParentID { get; set; }
+        public long ID { get; set; }
+        public long ParentID { get; set; }
     }
 }
